refactor(ui): added `CarouselGrid` component HBL-187 #time 30m
