+++ libs/ui/components/src/lib/CarouselGrid/index.ts
@@ -1,0 +1,3 @@
+import CarouselGrid from './CarouselGrid';
+
+export default CarouselGrid;
